Camera/Linux/requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
ultralytics>=8.0.0
numpy>=1.24.0
requests>=2.31.0
websocket-client>=1.8.0
Pillow>=9.5.0
websocket-client>=1.6.0
